python/hyperliquid-hypercore/pyproject.toml
[project]
name = "hyperliquid-hypercore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ecdsa>=0.19.1",
    "hyperliquid-python-sdk>=0.19.0",
    "path>=17.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
